Changes made
I made a few minor touch ups in different components such as; changing the routes in AppRouter so that SignUp is the main page then changed the order to SignUp, Login, and Profile. I also imported PrivateRoute to the AppRouter and added it to Profilepath to make it private. I took out the bullet point and changed it to a nav in AppRouter.js. In Login.js I added an alert if the username or password was typed incorrectly, but I might think about using Yup instead.