Login updates
I figured out on how to have when the user hits submit in the login, that they will be directed to the profile page. I also added some more styling. I will come back and test out some of the styled-components themes to help with styling. I also added the function to when the user hits submit, their login info disapears.